framework/fit/python/requirements.txt
PyYAML==6.0.1
flask==2.2.5
numpy==1.25.2
requests==2.31.0